feat(login/common): merge mixins and utils into single file
